requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
asyncpg==0.30.0
python-dotenv==1.0.1
google-genai==0.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pypdf==4.3.1
python-docx==1.1.2
ebooklib==0.18
lxml==5.3.0
alembic==1.14.1
